pyproject.toml
[tool.poetry]
name = "fintech"
version = "0.1.0"
description = ""
authors = ["Dmitry <dmitry.emelianenko@lyte.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.2"
djangorestframework-jsonapi = "^7.0.2"
uuid = "^1.30"
mysqlclient = "^2.2.4"
mypy = "^1.10.1"
flake8 = "^7.1.0"
black = "^24.4.2"
isort = "^5.13.2"
pytest = "^8.2.2"
pytest-factoryboy = "^2.7.0"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py311']
skip-string-normalization = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "FinTech.settings"

[tool.mypy]
ignore_missing_imports = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
warn_unused_ignores = true
warn_return_any = true
warn_no_return = false
warn_unreachable = true
strict_equality = true
warn_redundant_casts = true
exclude = [
    '.migrations.',
    'manage.py',
    'FinTech.settings.py'
]


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
skip_glob = ['venv']
combine_as_imports = true
combine_star = true
skip = [
    '.git',
    '.mypy_cache',
    '.ipython',
    '.venv',
]


[tool.flake8]
ignore = "D203,F403,F401,W503,W605"
exclude = [
    '.git',
    '.pytest_cache',
    '.idea',
    '.venv',
    '__pycache__',
]
max-complexity = 10
max-line-length = 120